Add API for creating a page
Since the title of a page is what is going to be used to link to them
from blocks the title parameter has to be unique. Rather than storing a
title as the exact id of the document I went with the approach of
slugifying the title. This makes it more url friendly and still can be
enforced as unique since the tranform is easy to operate.

To avoid possible confusion from users I went the route of having the
title be case sensitive. So "My First Note" and "my first note" are two
different notes and would be allowed.