Add creating block as a top level page element
Since a block must belong to a page and page ids are their titles which
also has to be unique it makes some sense to have the id of a block also
contain the page id. This gives more uniqueness in total and avoids
having to use long uuid's for id's.