Add CouchDB exceptions and refactor clients
Couch's API is really nice to work with and has good status codes for
different errors that can happen. I turned each case into an exception
that we can build and return from API calls. This makes it easier to
unpack values out of client calls.

Duplicating the usage of the new exceptions to all the clients was going
to be a lot of work. Since the clients aren't really doing that much
heavy lifting I decided to refactor to only have one client where all
the exception mapping happens. Callers into the client now need to build
urls on their own, but its not all that bad.