Add couch.view.sync mix task
This task will take views defined in `/priv/couch` and sync them up to
couchdb. This makes it so that views can be maintained in source control
and also have proper syntax highlighting when writing view map
functions.

Currently only map functions are supported, but once we need then a view
definition should be able to add a `reduce.js` and the script extended
to consume them.